Fix Q28_4

+++ src/ch/epfl/javelo/Q28_4.java
@@ -22,8 +22,7 @@ public final class Q28_4 {
      */
     public static double asDouble(int q28_4){
         //divide by 2^4 to obtain the decimal number representation in Q28_4 (decimal point before 4th bit)
-        //Is this correct? What about negative values?
-        return Math.scalb(q28_4, -4);
+        return Math.scalb((double)q28_4, -4);
     }
 
     /**
@@ -32,6 +31,6 @@ public final class Q28_4 {
      * @return (float) Q28.4 representation
      */
     public static float asFloat(int q28_4){
-        return Math.scalb(q28_4, -4);
+        return Math.scalb((float)q28_4, -4);
     }
 }
